lint(project): add missing return types

+++ src/resolvers/topic.resolver.ts
@@ -1,13 +1,14 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Parent, Query, ResolveField, Resolver } from '@nestjs/graphql';
-import { Topic } from '@prisma/client';
+import { Note, Topic, User } from '@prisma/client';
 
-import { User } from '@/common/decorators/request-user.decorator';
+import { User as UserDecorator } from '@/common/decorators/request-user.decorator';
 import { UuidInput } from '@/common/dtos/id.dto';
 import { TopicCreateInput, TopicFilterInput, TopicUpdateInput } from '@/common/dtos/topic.dto';
 import { authGuardian } from '@/common/guards/authentication/auth.guard';
+import { Paginated } from '@/common/interfaces/paginated.interface';
 import { NoteModel } from '@/common/models/note.model';
-import { TopicModel } from '@/common/models/topic.model';
+import { PaginatedTopicModel, TopicModel } from '@/common/models/topic.model';
 import { UserModel } from '@/common/models/user.model';
 import { NoteService } from '@/services/note.service';
 import { TopicService } from '@/services/topic.service';
@@ -23,39 +24,41 @@ export class TopicResolver {
 
   @Mutation(() => TopicModel)
   @UseGuards(authGuardian())
-  async createTopic(@Args('data') data: TopicCreateInput, @User('id') userId: string) {
+  createTopic(@Args('data') data: TopicCreateInput, @UserDecorator('id') userId: string): Promise<Topic> {
     return this.topicService.create({ ...data, userId });
   }
 
   @Query(() => TopicModel)
-  async getTopic(@Args('data') data: UuidInput) {
+  @UseGuards(authGuardian())
+  getTopic(@Args('data') data: UuidInput): Promise<Topic> {
     return this.topicService.getById(data.id);
   }
 
-  @Query(() => [TopicModel])
-  async getTopics(@Args('data', { nullable: true }) data?: TopicFilterInput) {
+  @Query(() => PaginatedTopicModel)
+  @UseGuards(authGuardian())
+  getTopics(@Args('data', { nullable: true }) data?: TopicFilterInput): Promise<Paginated<Topic>> {
     return this.topicService.getAll(data);
   }
 
   @Mutation(() => TopicModel)
   @UseGuards(authGuardian())
-  async updateTopic(@Args('data') data: TopicUpdateInput) {
-    return this.topicService.update(data);
+  updateTopic(@Args('data') data: TopicUpdateInput, @UserDecorator('id') userId: string): Promise<Topic> {
+    return this.topicService.update({ ...data, userId });
   }
 
   @Mutation(() => TopicModel)
   @UseGuards(authGuardian())
-  async deleteTopic(@Args('data') data: UuidInput) {
-    return this.topicService.delete(data);
+  deleteTopic(@Args('data') data: UuidInput, @UserDecorator('id') userId: string): Promise<Topic> {
+    return this.topicService.delete({ ...data, userId });
   }
 
   @ResolveField(() => NoteModel)
-  async lastNote(@Parent() topic: Topic) {
-    return this.noteService.getLastNoteByTopicId(topic.id);
+  lastNote(@Parent() topic: Topic, @UserDecorator('id') userId: string): Promise<Note> {
+    return this.noteService.getLastNoteByTopicId({ topicId: topic.id, userId });
   }
 
   @ResolveField(() => UserModel)
-  async user(@Parent() topic: Topic) {
+  user(@Parent() topic: Topic): Promise<User> {
     return this.userService.getById(topic.userId);
   }
 }
